+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/BasicAutonomous.java
@@ -2,91 +2,167 @@ package org.firstinspires.ftc.teamcode;
 
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
+import com.qualcomm.robotcore.hardware.CRServo;
 import com.qualcomm.robotcore.hardware.DcMotor;
-import com.qualcomm.robotcore.hardware.DcMotorController;
-import com.qualcomm.robotcore.hardware.DcMotorImpl;
-import com.qualcomm.robotcore.util.ElapsedTime;
+import com.qualcomm.robotcore.hardware.Servo;
 
 
 @Autonomous (name = "BasicAutonomous", group = "teamcode")
 
 
 public class BasicAutonomous extends LinearOpMode {
-    public DcMotor leftFront = null;
-    public DcMotor rightFront = null;
-    public DcMotor leftBack = null;
-    public DcMotor rightBack = null;
+    public DcMotor leftFrontDrive = null;
+    public DcMotor rightFrontDrive = null;
+    public DcMotor leftBackDrive = null;
+    public DcMotor rightBackDrive = null;
+    public DcMotor armMotor = null;
+    public CRServo intake      = null;
+    public Servo wrist       = null;
 
 
     @Override
     public void runOpMode() throws InterruptedException {
-        leftFront = hardwareMap.get(DcMotor.class, "left_front_drive");
-        rightFront = hardwareMap.get(DcMotor.class, "right_front_drive");
-        leftBack = hardwareMap.get(DcMotor.class, "left_back_drive");
-        rightBack = hardwareMap.get(DcMotor.class, "right_back_drive");
-        leftFront.setMode(DcMotor.RunMode.RUN_WITHOUT_ENCODER);
-        leftBack.setMode(DcMotor.RunMode.RUN_WITHOUT_ENCODER);
-        rightFront.setMode(DcMotor.RunMode.RUN_WITHOUT_ENCODER);
-        rightBack.setMode(DcMotor.RunMode.RUN_WITHOUT_ENCODER);
+        leftFrontDrive = hardwareMap.get(DcMotor.class, "left_front_drive");
+        rightFrontDrive = hardwareMap.get(DcMotor.class, "right_front_drive");
+        leftBackDrive = hardwareMap.get(DcMotor.class, "left_back_drive");
+        rightBackDrive = hardwareMap.get(DcMotor.class, "right_back_drive");
+        armMotor = hardwareMap.get(DcMotor.class, "arm_motor");
+        intake = hardwareMap.get(CRServo.class, "intake");
+        wrist = hardwareMap.get(Servo.class, "wrist");
 
-        leftFront.setDirection(DcMotor.Direction.REVERSE);
-        leftBack.setDirection(DcMotor.Direction.REVERSE);
-        rightFront.setDirection(DcMotor.Direction.FORWARD);
-        rightBack.setDirection(DcMotor.Direction.FORWARD);
+        leftFrontDrive.setMode(DcMotor.RunMode.RUN_WITHOUT_ENCODER);
+        leftBackDrive.setMode(DcMotor.RunMode.RUN_WITHOUT_ENCODER);
+        rightFrontDrive.setMode(DcMotor.RunMode.RUN_WITHOUT_ENCODER);
+        rightBackDrive.setMode(DcMotor.RunMode.RUN_WITHOUT_ENCODER);
+        armMotor.setMode(DcMotor.RunMode.RUN_WITHOUT_ENCODER);
+
+        leftFrontDrive.setDirection(DcMotor.Direction.FORWARD);
+        leftBackDrive.setDirection(DcMotor.Direction.FORWARD);
+        rightFrontDrive.setDirection(DcMotor.Direction.REVERSE);
+        rightBackDrive.setDirection(DcMotor.Direction.REVERSE);
+        armMotor.setDirection(DcMotor.Direction.FORWARD);
+        intake.setDirection(CRServo.Direction.FORWARD);
+        wrist.setDirection(Servo.Direction.FORWARD);
+
 
 
         waitForStart();
 
         while (opModeIsActive()) {
-           moveFoward(0.1, 4000);
-           stopMotor();
+
+
+            moveFoward(0.5, 298);
+            moveArm(0.25, 1372);
+            wrist.setPosition(0.5);
+            intake.setPower(1);
+            intake.setPower(0);
+            wrist.setPosition(1);
+            moveArm(0.25, -1372);
+            moveBack(0.5, 298);
+            moveRight(0.5, 1067);
+            moveFoward(0.5, 178);
+            moveLeft(0.5, 1067);
+            moveRight(0.5, 1067);
+            moveFoward(0.5, 223);
+            moveLeft(0.5, 1067);
+            moveBack(0.5, 401);
+            moveBack(0.5, 2133);
+            moveRight(0.5, 1066);
+            moveBack(0.5, 178);
+            moveLeft(0.5, 800);
+            moveRight(0.5, 267);
+            moveBack(0.5, 355);
+            moveLeft(0.5, 267);
+            moveFoward(0.5, 2311);
+            moveFoward(0.5, 355);
+            moveBack(0.5, 355);
+            moveBack(0.5,2133);
+            //parked :)
+
         }
 
 
     }
 
     public void moveFoward(double power, long time) {
-        leftFront.setPower(power);
-        leftBack.setPower(power);
-        rightFront.setPower(power);
-        rightBack.setPower(power);
-
+        leftFrontDrive.setPower(power);
+        leftBackDrive.setPower(power);
+        rightFrontDrive.setPower(power);
+        rightBackDrive.setPower(power);
         sleep(time);
+        stopMotor();
 
-
+    }
+    public void moveBack(double power, long time) {
+        leftFrontDrive.setPower(-power);
+        leftBackDrive.setPower(-power);
+        rightFrontDrive.setPower(-power);
+        rightBackDrive.setPower(-power);
+        sleep(time);
+        stopMotor();
 
     }
 
     public void turnRight(double power, long time) {
-        leftFront.setPower(power);
-        leftBack.setPower(power);
-        rightFront.setPower(-power);
-        rightBack.setPower(-power);
+        leftFrontDrive.setPower(power);
+        leftBackDrive.setPower(power);
+        rightFrontDrive.setPower(-power);
+        rightBackDrive.setPower(-power);
 
         sleep(time);
-
+        stopMotor();
 
     }
 
+    public void moveRight(double power, long time) {
+        leftFrontDrive.setPower(power);
+        leftBackDrive.setPower(-power);
+        rightFrontDrive.setPower(-power);
+        rightBackDrive.setPower(power);
+        sleep(time);
+        stopMotor();
+
+    }
+
+    public void moveLeft(double power, long time) {
+        leftFrontDrive.setPower(-power);
+        leftBackDrive.setPower(power);
+        rightFrontDrive.setPower(power);
+        rightBackDrive.setPower(-power);
+        sleep(time);
+        stopMotor();
+    }
 
     public void turnLeft(double power, long time) {
 
-        leftFront.setPower(power);
-        leftBack.setPower(power);
-        rightFront.setPower(-power);
-        rightBack.setPower(-power);
+        leftFrontDrive.setPower(-power);
+        leftBackDrive.setPower(-power);
+        rightFrontDrive.setPower(power);
+        rightBackDrive.setPower(power);
 
         sleep(time);
+        stopMotor();
+
+    }
+
+    public void moveArm(double power, long time) {
+
+        armMotor.setPower(power);
+
+        sleep(time);
+        stopMotor();
 
 
     }
+
     public void stopMotor(){
-        leftFront.setPower(0);
-        leftBack.setPower(0);
-        rightFront.setPower(0);
-        rightBack.setPower(0);
-
-
+        leftFrontDrive.setPower(0);
+        leftBackDrive.setPower(0);
+        rightFrontDrive.setPower(0);
+        rightBackDrive.setPower(0);
+        armMotor.setPower(0);
+        intake.setPower(0);
     }
-}
 
+
+}
